Command Block Development
Added CommandBlocks which are visual representations of commands. Not done yet, and not tested either.

Also, made slight edits to Command and ScriptStruct. Command was moved to the prefab package so it could be imported to other scripts (reflecting this was the change in ScriptStruct). Also switched hashtable.add for hashtable.put