Vertical Sorting Pane and Drag-And-Drop Realignment
Added more functionality to VerticalSortingPane (VSP), specifically the (theoretical) ability for VSP items to correct their position if something is being dragged over them.

Still needs testing and some more functionality, like adding and removing  items and the VSP adjusting to that.

Also added 2 custom events to handle CommandBlocks interacting with the VSP without building in a dependency on checking for parent types.