Command Flow updating through VSP
Command Flow 'should' now be synced with the VSP representation of the Command Blocks.

I say 'should', but it totally isn't. I haven't tested it exporting anything and it throws a lot of errors when I move things around, but the VSP itself still works so I'll fix it later because my hands hurt a ton right now.