Added Over/Undershoot handlers
The overshoot/undershoot handlers are temporary handlers to take care of command blocks that relocate off of the list. This shouldn't be possible, but it happens so much that I'm just putting in a handler for right now.

Also removed old debug code and turned the sidebar back into a VBox for now